reorg organization methods, add new method, update linter/CI
