+++ .buildkite/pipeline.yml
@@ -10,7 +10,7 @@ steps:
     - docker#v5.11.0:
         environment:
           - GOFLAGS=-buildvcs=false
-        image: "registry.hub.docker.com/golangci/golangci-lint:v1.57-alpine"
+        image: "registry.hub.docker.com/golangci/golangci-lint:v1.60.3-alpine"
         command: ["golangci-lint", "run", "-v", "--timeout", "5m"]
 
 - label: ":test_tube: test"
